refactor(cli): move Spica trace adapter out of main
