+++ tests/unit/cli/test_cli_workflow.py
@@ -17,18 +17,20 @@ import pytest
 import yaml
 
 from aiconfigurator.cli.main import (
-    _build_spica_trace_result_bundle,
-    _build_spica_trace_search_space,
     _execute_tasks,
     _resolve_cli_log_level,
-    _save_spica_trace_artifacts,
-    _spica_candidates_to_result_df,
     build_default_tasks,
     build_experiment_tasks,
     configure_parser,
 )
 from aiconfigurator.cli.main import main as cli_main
 from aiconfigurator.cli.report_and_save import _apply_inclusive_tpot
+from aiconfigurator.cli.spica.trace_adapter import (
+    _build_spica_trace_result_bundle,
+    _build_spica_trace_search_space,
+    _save_spica_trace_artifacts,
+    _spica_candidates_to_result_df,
+)
 from aiconfigurator.sdk.errors import NoFeasibleConfigError
 
 pytestmark = pytest.mark.unit
@@ -165,7 +167,7 @@ class TestCLIIntegration:
         mock_builder.assert_called_once()
         mock_execute.assert_called_once_with({"agg": mock_task_config}, mode, top_n=5)
 
-    @patch("aiconfigurator.cli.main._run_spica_trace_default")
+    @patch("aiconfigurator.cli.main.run_spica_trace_default")
     @patch("aiconfigurator.cli.main._execute_tasks")
     @patch("aiconfigurator.cli.main.build_default_tasks")
     def test_cli_default_trace_path_dispatches_to_spica(
